+++ paper/paper.bib
@@ -28,7 +28,7 @@
   year={2014},
 }
 
-@article{expyriment,
+@article{neuropsydia,
   title={Neuropsydia.py: A Python module for creating experiments, tasks and questionnaires},
   journal={Journal of Open Source Software},
   author={Makowski, Dominique and Dutriaux, Léo},
