### 2018-10-23 b845cb3
better logging routine for key records
