+++ ElGateau/ElGateau.py
@@ -18,7 +18,7 @@ __author__ = "Christopher Madan"
 __copyright__ = "Copyright 2017-2018, Christopher Madan"
 
 __license__ = "MIT"
-__version__ = "0.9.1"
+__version__ = "0.9.2"
 __maintainer__ = "Christopher Madan"
 __email__ = "christopher.madan@nottingham.ac.uk"
 __status__ = "Production"
@@ -539,7 +539,8 @@ class ElGateau(object):
         
         # if logging, make a record
         if hasattr(self,'Log'):
-            self.Log.record_display(key,icon)
+            # need to remap key for log to match intuition
+            self.Log.record_display(self.key_remap(key),icon)
 
     ########################################
     #
@@ -648,7 +649,8 @@ class ElGateau(object):
 
         # if logging, make a record
         if hasattr(self,'Log'):
-            self.Log.record_press(button,response_time)
+            # need to remap key for log to match intuition
+            self.Log.record_press(self.key_remap(button),response_time)
 
         return (button, response_time)
 
