+++ README.md
@@ -68,8 +68,7 @@ pip install ...
 ### Disclaimer
 
 This project is not associated with Elgato Systems GmbH. 
-Emoji artwork is provided by EmojiOne (v2.3) and is licensed under CC-BY 4.0.
-(pixel fonts?)
+Emoji artwork is provided by EmojiOne (v2.3) and is licensed under CC-BY 4.0. See associated licences accompanying each font.
 
 
 ### Acknowledgements
