### 2018-10-30 #1046 @Kevin-Mattheus-Moerman
@JonathanReardon would you be able to review this submission for [JOSS](http://joss.theoj.org/)? The [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) takes place on GitHub and focuses on [the software](https://github.com/cMadan/ElGateau) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.01046/joss.01046/10.21105.joss.01046.pdf). Let me know if you have questions.
