### 2018-10-30 #1046 @Kevin-Mattheus-Moerman
@cMadan thanks for this submission. Since this submission features a software to interact with a specific type of hardware device, can you say a couple of words about:

1. What aspects of this work can be evaluated/reviewed without the device?
2. Is the device needed to properly review this work?
3. If so, are you able to lend the device to a reviewer (we have made such arrangements for other works in the past)?
4. Have you programmed a virtual version of the hardware device i.e. a simulation environment so some of the functionality of the software can be tested in the absence of the device?
