### 2018-10-30 #1046 @cMadan
Thanks for getting things started, @Kevin-Mattheus-Moerman!

> What aspects of this work can be evaluated/reviewed without the device?

The toolbox includes a developer mode that simulates the device i/o, that functions within Jupyter Notebook.

> Is the device needed to properly review this work? If so, are you able to lend the device to a reviewer (we have made such arrangements for other works in the past)?

The device isn't needed, since there is the developer mode. I also could demonstrate the device remotely (e.g., over Skype), doing a live demo of the device interaction on the device itself.

> Have you programmed a virtual version of the hardware device i.e. a simulation environment so some of the functionality of the software can be tested in the absence of the device?

Yes, there is a simulation mode built-in the code.
