### 2018-11-08 #1070 @whedon
[REVIEW]: ElGateau: A Library for Using the Elgato Stream Deck for Experimental Psychology Research
**Submitting author:** @cMadan (<a href="http://orcid.org/0000-0003-3228-6501">Christopher R Madan</a>)
**Repository:** <a href="https://github.com/cMadan/ElGateau" target ="_blank">https://github.com/cMadan/ElGateau</a>
**Version:** v1.0.1
