### 2018-11-20 #1070 @JonathanReardon
Apologies for the delay in getting back to you, I will be able to work on the review tomorrow, Thursday, and Friday.
