### 2018-11-24 #1070 @JonathanReardon
Hi all, I have now completed the review. The library works as described in the documentation and I experienced no problems during all parts of testing. @cMadan, it might be a good idea to add 'Ubuntu 18.10 (Linux)' to the 'tested on..' statement in the README.md file, since this was the operating system that I tested the library on. Thanks all.
