### 2018-11-25 f4dee03
setup.py for pip install
