### 2018-11-25 f9d7209
Increment to v. 1.0.0
