### 2018-11-26 8595b96
Better path handling
