### 2018-11-26 ab81ed0
Fixed pip install issues
