+++ ElGateau/__init__.py
@@ -1,0 +1,1 @@
+from ElGateau.ElGateau import *
