### 2018-11-26 f1d7108
fix for external icons and fonts
