### 2018-11-26 d60d0f4
Update device image to be photo of demo
