### 2018-11-26 #1070 @cMadan
I changed the photo in the paper (and readme) to be a photo of the device running one of the demos, rather than a photo I had found online.
