### 2018-11-26 #1070 @Kevin-Mattheus-Moerman
@arfon I recommend we accept this submission. The version is now at  v1.0.1. The DOI of the archived version is: 10.5281/zenodo.1525298.
